Fixed warning. JIRA: CXX-7834
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@73805 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/test/objmgr/id_unit_test_bad.cpp
@@ -83,7 +83,7 @@ static CRandom s_Random;
 template<class I>
 void s_RandomShuffle(I iter1, I iter2)
 {
-    for ( size_t s = distance(iter1, iter2); s > 1; --s, ++iter1 ) {
+    for ( int s = int(distance(iter1, iter2)); s > 1; --s, ++iter1 ) {
         swap(*iter1, *next(iter1, s_Random.GetRandIndex(s)));
     }
 }
